Refactored to reduce paramerts for grip form
